ml-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
supabase==1.1.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
joblib==1.3.2
python-multipart==0.0.6
cors==1.0.1